+++ picopt.sh
@@ -1,3 +1,4 @@
-#!/bin/sh
+#!/usr/bin/env bash
 # Run picopt in development
+export PYTHONDEVMODE=1
 uv run ./picopt.py "$@"